+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Fix [#609](https://github.com/squint-cljs/squint/issues/609): make `remove` return a transducer when no collection is provided
 - Fix [#611](https://github.com/squint-cljs/squint/issues/611): Implement the `set?` function
+- Fix [#613](https://github.com/squint-cljs/squint/issues/613): Optimize `aset`
 
 ## v0.8.134 (2024-02-11)